+++ demo-app/resources/js/Pages/ShowUser.jsx
@@ -1,13 +1,24 @@
 import { ModalLink } from '@inertiaui/modal-react';
 import Container from './Container';
+import * as InertiaReact from '@inertiajs/react';
 
-export default function UserProfile({ user }) {
+export default function UserProfile({ user, deferred }) {
     return (
         <Container>
             <div className="">
                 <p className="text-2xl font-medium text-gray-900">{user.name}</p>
                 <p className="text-xl text-gray-500">{user.email}</p>
             </div>
+
+            {
+                InertiaReact.Deferred ?
+                    <InertiaReact.Deferred data="deferred" fallback={<div>Loading...</div>}>
+                        <p dusk="deferred">
+                            {deferred}
+                        </p>
+                    </InertiaReact.Deferred> : <p dusk="deferred"> No Deferred Component</p>
+            }
+
             <ModalLink
                 navigate
                 as="button"
@@ -17,6 +28,6 @@ export default function UserProfile({ user }) {
             >
                 Edit
             </ModalLink>
-        </Container>
+        </Container >
     );
 };