+++ demo-app/tests/ModalTestCase.php
@@ -2,6 +2,8 @@
 
 namespace Tests;
 
+use InertiaUI\Modal\Support;
+
 trait ModalTestCase
 {
     public static function booleanProvider(): array
@@ -10,5 +12,10 @@ trait ModalTestCase
             'true' => [true],
             'false' => [false],
         ];
+    }
+
+    public function isInertiaV2(): bool
+    {
+        return Support::isInertiaV2();
     }
 }