+++ src/Support.php
@@ -1,0 +1,18 @@
+<?php
+
+declare(strict_types=1);
+
+namespace InertiaUI\Modal;
+
+class Support
+{
+    public static function isInertiaV1(): bool
+    {
+        return ! class_exists(\Inertia\DeferProp::class);
+    }
+
+    public static function isInertiaV2(): bool
+    {
+        return class_exists(\Inertia\DeferProp::class);
+    }
+}